interview_app/requirements.txt
Flask-SQLAlchemy 


# python command to create database

# python -c "from app import create_app, db; app = create_app(); with app.app_context(): db.create_all()"